found issue with file selection.  Fixed

+++ models/JumpNumber.cs
@@ -19,7 +19,7 @@ namespace IntelReader.models
     public class LogFileInfo
     {
         public long length;
-        public Int32 fileDate;
+        public long fileDate;
 //        public long lines;
         public bool firstCheck;
         public string fullName;
